backend/django/requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1, <0.16.0
requests>=2.31.0,<2.32.0
django-cors-headers==3.8.0
boto3>=1.34,<1.35
uwsgi>=2.0.19,<2.1
drf_yasg== 1.21.7
uuid==1.30
jsonschema